secure-dev-training/requirements.txt
flask-sqlalchemy
flask-migrate
flask-login
bcrypt
flask-wtf
wtforms-json